framework/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-shopman"
version = "0.1.0"
description = "Django Shopman — omnichannel commerce framework for Django"
license = "MIT"
authors = [
    {name = "Pablo Valentini", email = "pablondrina@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "Django>=5.2,<6.0",
    "djangorestframework>=3.15,<4.0",
    "drf-spectacular>=0.28,<1.0",
    "shopman-utils>=0.3.0",
    "shopman-omniman>=0.1.0",
    "shopman-stockman>=0.3.0",
    "shopman-craftsman>=0.3.0",
    "shopman-offerman>=0.3.0",
    "shopman-guestman>=0.1.0",
    "shopman-doorman>=0.1.0",
    "shopman-payman>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["shopman*", "project*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings"
pythonpath = ["."]

[tool.coverage.run]
source = ["shopman"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*_test_settings*",
    "manage.py",
    "project/wsgi.py",
    "project/asgi.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 0
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__",
]

[tool.ruff]
line-length = 120
target-version = "py311"
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]
